refactor: move more wallet-related methods to wallet.rs
