+++ src/dfx/src/lib/migrate.rs
@@ -10,6 +10,7 @@ use ic_utils::{
 };
 use itertools::Itertools;
 
+use crate::lib::identity::wallet::wallet_canister_id;
 use crate::lib::operations::canister::install_wallet;
 
 use super::{
@@ -28,7 +29,7 @@ pub async fn migrate(env: &dyn Environment, network: &NetworkDescriptor, fix: bo
     let mut mgr = env.new_identity_manager()?;
     let ident = mgr.instantiate_selected_identity(env.get_logger())?;
     let mut did_migrate = false;
-    let wallet = if let Some(principal) = Identity::wallet_canister_id(network, ident.name())? {
+    let wallet = if let Some(principal) = wallet_canister_id(network, ident.name())? {
         principal
     } else {
         bail!("No wallet found; nothing to do");
